backend/ml_models/pest_detection/requirements.txt
tensorflow>=2.15.0
keras>=2.15.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
tqdm>=4.65.0
jupyter>=1.0.0
